+++ lib/io.dart
@@ -8,12 +8,13 @@ import "dart:io";
 import "dart:async";
 import "dart:collection" show LinkedHashMap;
 import "dart:json" as Json;
+import "dart:convert";
 import "package:meta/meta.dart";
 
 import "async.dart";
+import "convert.dart";
 
 part "src/io/http_wrapper.dart";
 part "src/io/iosink_wrapper.dart";
-part "src/io/io_util.dart";
 part "src/io/http_util.dart";
 part "src/io/gzip_util.dart";