+++ test/io_test.dart
@@ -10,20 +10,6 @@ import "package:rikulo_commons/io.dart";
 
 main() {
   group("io tests", () {
-    test("decode/encode", () {
-      final str = "abcdefg";
-      expect(decodeString(encodeString(str)), str);
-    });
-
-    test("stream to json", () {
-      final val = {"abc": 123, "foo": ["this", "is", 200]};
-      final list = encodeString(stringify(val));
-      return IOUtil.readAsJson(new Stream.fromIterable([list]))
-      .then((got) {
-        expect(got, val);
-      });
-    });
-
     test("gzip test", () {
       final buf = new StringBuffer();
       for (int i = 100; --i >= 0;)